@jbeans0312 Moved around some buttons, added contextual visibility for some buttons, moved the concentration selecter out of semester and into plan.
