+++ src/components/SemesterView.tsx
@@ -4,8 +4,6 @@ import { Course } from "../interfaces/course";
 import DeleteIcon from "@mui/icons-material/Delete";
 import Button from "@mui/material/Button";
 import { CourseList } from "./CourseList";
-import { Form } from "react-bootstrap";
-import { Plan } from "../interfaces/plan";
 import { useDrop } from "react-dnd";
 
 export function SemesterView({
@@ -21,33 +19,6 @@ export function SemesterView({
     pool: Course[];
     setPool: (newPool: Course[]) => void;
 }): JSX.Element {
-    function checkCourse(thisSemester: Semester, Query: Course): boolean {
-        return thisSemester.courses.indexOf(Query) !== -1;
-    }
-    function checkPlan(thisPlan: Plan, Query: Course): boolean {
-        const goodSemester: Semester[] = thisPlan.semesters.filter(
-            (aSemester: Semester): boolean => checkCourse(aSemester, Query)
-        );
-        return goodSemester.length > 0;
-    }
-    function checkConcentration(aPlan: Plan, courseList: Course[]): boolean {
-        const missingCourses: Course[] = courseList.filter(
-            (aCourse: Course): boolean => checkPlan(aPlan, aCourse) === false
-        );
-        return missingCourses.length === 0;
-    }
-    function needTheseCourse(aPlan: Plan, courseList: Course[]): Course[] {
-        const missingCourses: Course[] = courseList.filter(
-            (aCourse: Course): boolean => checkPlan(aPlan, aCourse) === false
-        );
-        return missingCourses;
-    }
-    const [concentrationPicked, setconcentrationPicked] =
-        useState<string>("AI");
-
-    function updateConcentration(event: React.ChangeEvent<HTMLSelectElement>) {
-        setconcentrationPicked(event.target.value);
-    }
     const [courses, setCourses] = useState<Course[]>(semester.courses);
     /** Use state and implementation of drag functionality, linked with DraggableCourse.tsx and CourseFinder.tsx */
     /** Calls updateCourses when a new course is dropped into a semester */
@@ -126,49 +97,6 @@ export function SemesterView({
                     Delete Semester
                 </Button>
             ) : null}
-            <div>
-                <Form.Group controlId="userConcentration">
-                    <Form.Label>What is your concentration?</Form.Label>
-                    <Form.Select
-                        value={concentrationPicked}
-                        onChange={updateConcentration}
-                    >
-                        <option value="AI">AI</option>
-                        <option value="Bioinformatics"> Bioinformatics </option>
-                        <option value="Cybersecurity"> Cybersecurity </option>
-                        <option value="DataScience"> Data Science </option>
-                        <option value="High Performance Computing">
-                            High Performance and Computing
-                        </option>
-                        <option value="Systems and Networks">
-                            Systems and Networks
-                        </option>
-                        <option value="Theory and Computation">
-                            Theory and Computation
-                        </option>
-                    </Form.Select>
-                </Form.Group>
-                The users concentration {concentrationPicked}.
-                <div>
-                    <span>
-                        The user has passed Concetration requirement.
-                        {checkConcentration}
-                        The user needs these courses still to pass Concentration
-                        requirement.
-                        {needTheseCourse}
-                        <span>The data should be checked here!</span>
-                    </span>
-                </div>
-            </div>
-            <Button
-                startIcon={<DeleteIcon />}
-                variant="outlined"
-                color="error"
-                className="m-2"
-                onClick={() => deleteSemester(semester.year, semester.season)}
-            >
-                Delete Semester
-            </Button>
             {isOver && <div>Insert Course!</div>}
         </div>
     );
